Extract a snackBarPadding Modifier
